SQL requests for selecting user by email and creating new user
