+++ src/main/java/dao/h2/H2UserDAO.java
@@ -9,6 +9,11 @@ import java.util.List;
 public class H2UserDAO implements UserDAO{
     private final ConnectionPool connectionPool;
 
+    private static final String SELECT_USER_BY_EMAIL_SQL =
+            "SELECT user_id, password, first_name, last_name, role from USERS WHERE email = ?";
+    private static final String CREATE_USER_SQL =
+            "INSERT INTO USERS VALUES(email, password, first_name, last_name, role) WHERE (?,?,?,?,?)";
+
     public H2UserDAO(ConnectionPool connectionPool) {
         this.connectionPool = connectionPool;
     }
