--- src/main/java/dao/implementation/SQLRequestConstants.java
@@ -1,4 +1,0 @@
-package dao.implementation;
-
-public class SQLRequestConstants {
-}
